--- a/lib/index.js
+++ b/lib/index.js
@@ -181,6 +181,100 @@ function defaultContext() {
 };
 
 /**
+ * A batch consists of 1 or more message parts with their flags that need to be sent as one unit
+ */
+
+function OutBatch() {
+  this.content = [];      // buf, flags, buf, flags, ...
+  this.cbs = [];          // callbacks
+  this.isClosed = false;  // true if the last message does not have SNDMORE in its flags, false otherwise
+  this.next = null;       // next batch (for linked list of batches)
+}
+
+OutBatch.prototype.append = function (buf, flags, cb) {
+  if (!Buffer.isBuffer(buf)) {
+    buf = new Buffer(String(buf), 'utf8');
+  }
+
+  this.content.push(buf, flags);
+
+  if (cb) {
+    this.cbs.push(cb);
+  }
+
+  if ((flags & zmq.SNDMORE) === 0) {
+    this.isClosed = true;
+  }
+};
+
+OutBatch.prototype.invokeError = function (socket, error) {
+  var returned = false;
+  for (var i = 0; i < this.cbs.length; i += 1) {
+    this.cbs[i].call(socket, error);
+    returned = true;
+  }
+
+  if (!returned) {
+    throw error;
+  }
+};
+
+OutBatch.prototype.invokeSent = function (socket) {
+  for (var i = 0; i < this.cbs.length; i += 1) {
+    this.cbs[i].call(socket);
+  }
+};
+
+
+function BatchList() {
+  this.firstBatch = null;
+  this.lastBatch = null;
+  this.length = 0;
+}
+
+BatchList.prototype.canSend = function () {
+  return this.firstBatch ? this.firstBatch.isClosed : false;
+};
+
+BatchList.prototype.append = function (buf, flags, cb) {
+  var batch = this.lastBatch;
+
+  if (!batch || batch.isClosed) {
+    batch = new OutBatch();
+
+    if (this.lastBatch) {
+      this.lastBatch.next = batch;
+    }
+
+    this.lastBatch = batch;
+
+    if (!this.firstBatch) {
+      this.firstBatch = batch;
+    }
+
+    this.length += 1;
+  }
+
+  batch.append(buf, flags, cb);
+};
+
+BatchList.prototype.fetch = function () {
+  var batch = this.firstBatch;
+  if (batch && batch.isClosed) {
+    this.firstBatch = batch.next;
+    this.length -= 1;
+    return batch;
+  }
+  return undefined;
+};
+
+BatchList.prototype.restore = function (batch) {
+  this.firstBatch = batch;
+  this.length += 1;
+};
+
+
+/**
  * Create a new socket of the given `type`.
  *
  * @constructor
@@ -195,20 +289,14 @@ exports.Socket = function (type) {
   this.type = type;
   this._zmq = new zmq.SocketBinding(defaultContext(), types[type]);
   this._paused = false;
+  this._isFlushingReads = false;
+  this._isFlushingWrites = false;
+  this._outgoing = new BatchList();
 
-  this._zmq.onReady = function() {
-    if(self._paused) return;
-
-    try {
-      self._flush();
-    } catch (err) {
-      self.emit('error', err);
-    } finally {
-      self._zmq.pending = self._outgoing.length;
-    }
-  }
-
-  this._outgoing = [];
+  this._zmq.onReady = function(readable, writable) {
+    if (readable) self._flushReads();
+    if (writable) self._flushWrites();
+  };
 };
 
 /**
@@ -218,8 +306,9 @@ exports.Socket = function (type) {
 util.inherits(Socket, EventEmitter);
 
 /**
- * Set pull socket to pause mode
+ * Set socket to pause mode
  * no data will be emit until resume() is called
+ * all send() calls will be queued
  *
  * @api public
  */
@@ -229,14 +318,15 @@ Socket.prototype.pause = function() {
 }
 
 /**
- * Set a pull back to normal work mode
+ * Set a socket back to normal work mode
  *
  * @api public
  */
 
 Socket.prototype.resume = function() {
   this._paused = false;
-  this._zmq.onReady();
+  this._flushReads();
+  this._flushWrites();
 }
 
 Socket.prototype.read = function() {
@@ -258,6 +348,7 @@ Socket.prototype.read = function() {
 
   return null;
 }
+
 
 /**
  * Set `opt` to `val`.
@@ -314,17 +405,16 @@ Object.keys(opts).forEach(function(name){
 
 Socket.prototype.bind = function(addr, cb) {
   var self = this;
-  self._zmq.bind(addr, function(err) {
-    try {
-      self._flush();
-      self.emit('bind', addr);
-      cb && cb(err);
-    } catch (err) {
-      self.emit('error', err);
-    } finally {
-      self._zmq.pending = self._outgoing.length;
+  this._zmq.bind(addr, function(err) {
+    if (err) {
+      return cb && cb(err);
     }
 
+    self._flushReads();
+    self._flushWrites();
+
+    self.emit('bind', addr);
+    cb && cb();
   });
   return this;
 };
@@ -338,11 +428,7 @@ Socket.prototype.bind = function(addr, cb) {
  */
 
 Socket.prototype.bindSync = function(addr) {
-  try {
-    this._zmq.bindSync(addr);
-  } catch (err) {
-    throw err;
-  }
+  this._zmq.bindSync(addr);
 
   return this;
 };
@@ -361,17 +447,18 @@ Socket.prototype.bindSync = function(addr) {
 Socket.prototype.unbind = function(addr, cb) {
   if (zmq.ZMQ_CAN_UNBIND) {
     var self = this;
-    self._zmq.unbind(addr, function(err) {
-      try {
-        self._flush();
-        self.emit('unbind', addr);
-        cb && cb(err);
-      } catch (err) {
-        self.emit('error', err);
-      } finally {
-        self._zmq.pending = self._outgoing.length;
+    this._zmq.unbind(addr, function(err) {
+      if (err) {
+        return cb && cb(err);
       }
+      self.emit('unbind', addr);
+
+      self._flushReads();
+      self._flushWrites();
+      cb && cb();
     });
+  } else {
+    cb && cb();
   }
   return this;
 };
@@ -386,11 +473,7 @@ Socket.prototype.unbind = function(addr, cb) {
 
 Socket.prototype.unbindSync = function(addr) {
   if (zmq.ZMQ_CAN_UNBIND) {
-    try {
-      this._zmq.unbindSync(addr);
-    } catch (err) {
-      throw err;
-    }
+    this._zmq.unbindSync(addr);
   }
   return this;
 }
@@ -493,6 +576,7 @@ Socket.prototype.unsubscribe = function(filter) {
   return this;
 };
 
+
 /**
  * Send the given `msg`.
  *
@@ -508,124 +592,105 @@ Socket.prototype.send = function(msg, flags, cb) {
 
   if (Array.isArray(msg)) {
     for (var i = 0, len = msg.length; i < len; i++) {
-      var part = msg[i];
       var isLast = i === len - 1;
+      var msgFlags = isLast ? flags : flags | zmq.ZMQ_SNDMORE;
+      var callback = isLast ? cb : undefined;
 
-      if (!Buffer.isBuffer(part)) {
-        part = new Buffer(String(part), 'utf8');
-      }
-
-      this._outgoing.push([part, isLast ? flags : zmq.ZMQ_SNDMORE, isLast ? cb : undefined]);
+      this._outgoing.append(msg[i], msgFlags, callback);
     }
   } else {
-    if (!Buffer.isBuffer(msg)) {
-      msg = new Buffer(String(msg), 'utf8');
-    }
-
-    this._outgoing.push([msg, flags, cb]);
+    this._outgoing.append(msg, flags, cb);
   }
 
-  if (!(flags & zmq.ZMQ_SNDMORE)) {
-    try {
-      this._flush();
-    } catch (err) {
-      this.emit('error', err);
-    } finally {
-      this._zmq.pending = this._outgoing.length;
-    }
+  if (this._outgoing.canSend()) {
+    this._zmq.pending = true;
+    this._flushWrites();
+  } else {
+    this._zmq.pending = false;
   }
 
   return this;
 };
 
 
-// The workhorse that does actual send and receive operations.
-// This helper is called from `send` above, and in response to
-// the watcher noticing the signaller fd is readable.
-Socket.prototype._flush = function() {
-  var flags, args, emitArgs, errorSent;
-
-  // Don't allow recursive flush invocation as it can lead to stack
-  // exhaustion and write starvation
-  if (this._flushing) return;
-
-  this._flushing = true;
-  while (true) {
-    // If an Async thread that may call a zmq function is in flight
-    // do not call any zmq functions.
-    if (this._zmq.state !== zmq.STATE_READY) {
-      this._flushing = false;
-      return;
-    }
-
-    flags = this._zmq.getsockopt(zmq.ZMQ_EVENTS);
-
-    if (!this._outgoing.length) flags &= ~zmq.ZMQ_POLLOUT;
-    if (!flags) {
-      this._flushing = false;
-      return;
-    };
-
-    var emitArgs = this.read();
-    if (emitArgs) {
-      emitArgs.unshift('message');
-      // Handle received message immediately to prevent memory leak in driver
-      this.emit.apply(this, emitArgs);
-    }
-
-    if (this._zmq.state !== zmq.STATE_READY) {
-      this._flushing = false;
-      return;
-    }
-
-    // We send as much as possible in one burst so that we don't
-    // starve sends if we receive more than one message for each
-    // one sent.
-    while (this._outgoing.length) {
-      flags = this._zmq.getsockopt(zmq.ZMQ_EVENTS);
-
-      if (!(flags & zmq.ZMQ_POLLOUT)) {
-        this._flushing = false;
-        return;
-      }
-
-      args = this._outgoing.shift();
-
-      try {
-        this._zmq.send(args[0], args[1]);
-
-        if (args[2]) {
-          args[2].call(this);
-        }
-      } catch (sendError) {
-        errorSent = false;
-
-        if (args[2]) {
-          errorSent = true;
-          args[2].call(this, sendError);
-        }
-
-        // If more chunks were to follow, we need to drop all of them.
-        // This loop will pull off the items up until and including
-        // the first item that is not flagged SNDMORE.
-
-        while (args && (args[1] & zmq.ZMQ_SNDMORE)) {
-          args = this._outgoing.shift();
-
-          if (args && args[2]) {
-            errorSent = true;
-            args[2].call(this, sendError);
-          }
-        }
-
-        if (!errorSent) {
-          this._flushing = false;
-          throw sendError;
-        }
-      }
-    }
+Socket.prototype._flushRead = function () {
+  var message = this._zmq.readv(); // can throw
+  if (!message) {
+    return false;
   }
-  this._flushing = false;
+
+  // Handle received message immediately to prevent memory leak in driver
+  if (message.length === 1) {
+    // hot path
+    this.emit('message', message[0]);
+  } else {
+    this.emit.apply(this, ['message'].concat(message));
+  }
+  return true;
+};
+
+Socket.prototype._flushWrite = function () {
+  var batch = this._outgoing.fetch();
+  if (!batch) {
+    this._zmq.pending = false;
+    return false;
+  }
+
+  try {
+    if (this._zmq.sendv(batch.content)) {
+      this._zmq.pending = this._outgoing.canSend();
+      batch.invokeSent(this);
+      return true;
+    }
+
+    this._outgoing.restore(batch);
+    return false;
+  } catch (sendError) {
+    this._zmq.pending = this._outgoing.canSend();
+    batch.invokeError(this, sendError); // can throw
+    return false;
+  }
+};
+
+
+Socket.prototype._flushReads = function() {
+  if (this._paused || this._isFlushingReads) return;
+
+  this._isFlushingReads = true;
+
+  var received;
+
+  do {
+    try {
+      received = this._flushRead();
+    } catch (error) {
+      this._isFlushingReads = false;
+      this.emit('error', error); // can throw
+      return;
+    }
+  } while (received);
+
+  this._isFlushingReads = false;
+};
+
+Socket.prototype._flushWrites = function() {
+  if (this._paused || this._isFlushingWrites) return;
+
+  this._isFlushingWrites = true;
+
+  var sent;
+
+  do {
+    try {
+      sent = this._flushWrite();
+    } catch (error) {
+      this._isFlushingWrites = false;
+      this.emit('error', error); // can throw
+      return;
+    }
+  } while (sent);
+
+  this._isFlushingWrites = false;
 };
 
 /**