--- a/test/socket.req-rep.js
+++ b/test/socket.req-rep.js
@@ -1,6 +1,5 @@
 var zmq = require('..')
-  , should = require('should')
-  , semver = require('semver');
+  , should = require('should');
 
 describe('socket.req-rep', function(){
   it('should support req-rep', function(done){
@@ -13,8 +12,9 @@ describe('socket.req-rep', function(){
       rep.send('world');
     });
 
-    rep.bind('inproc://stuff', function(){
-      req.connect('inproc://stuff');
+    rep.bind('inproc://stuffreqrep', function (error) {
+      if (error) throw error;
+      req.connect('inproc://stuffreqrep');
       req.send('hello');
       req.on('message', function(msg){
         msg.should.be.an.instanceof(Buffer);
@@ -40,7 +40,8 @@ describe('socket.req-rep', function(){
           rep.send('world');
         });
 
-        rep.bind('inproc://' + n, function(){
+        rep.bind('inproc://' + n, function (error) {
+          if (error) throw error;
           req.connect('inproc://' + n);
           req.send('hello');
           req.on('message', function(msg){