… test/socket.stream.js
… test/socket.unbind.js
… test/socket.zap.js
… test/zmq_proxy.push-pull.js
… test/zmq_proxy.router-dealer.js
… test/zmq_proxy.xpub-xsub.js
… test/zmq_proxy.xrep-xreq.js