This PR almost triples the throughput on the perf test. Buffers are *still* being copied when sent, so there's still room for improvement, but I would say this is a nice start. Review very welcome (cc @kkoopa @reqshark, @gtamas), please be rigorous.

Benchmark before:

```
$ make perf
node perf/local_lat.js tcp://127.0.0.1:5555 1 100000& node perf/remote_lat.js tcp://127.0.0.1:5555 1 100000
started receiving
message size: 1 [B]
roundtrip count: 100000
mean latency: 0.0733293089 [msecs]
overall time: 14 secs and 665861780 nanoseconds

node perf/local_thr.js tcp://127.0.0.1:5556 1 100000& node perf/remote_thr.js tcp://127.0.0.1:5556 1 100000
started receiving
message size: 1 [B]
message count: 100000
mean throughput: 40431 [msg/s]
mean throughput: 0 [Mbit/s]
overall time: 2 secs and 473361003 nanoseconds
```

Benchmark after:

```
$ make perf
node perf/local_lat.js tcp://127.0.0.1:5555 1 100000& node perf/remote_lat.js tcp://127.0.0.1:5555 1 100000
started receiving
message size: 1 [B]
roundtrip count: 100000
mean latency: 0.07873620208 [msecs]
overall time: 15 secs and 747240416 nanoseconds

node perf/local_thr.js tcp://127.0.0.1:5556 1 100000& node perf/remote_thr.js tcp://127.0.0.1:5556 1 100000
started receiving
message size: 1 [B]
message count: 100000
mean throughput: 115861 [msg/s]
mean throughput: 1 [Mbit/s]
overall time: 0 secs and 863099612 nanoseconds
```

Remaining tasks for this PR (post-review, when all is well):

- Remove old code from C++ (unused API replaced by this PR).

Remaining tasks post-merge:

- Code style unification.
- Deprecate ZMQ 2 support
- Release
- Remove ZMQ 2 support
- Deprecate ZMQ 3 support
- Release
- Remove ZMQ 3 support
- Avoid buffer copying when sending messages (GC challenge).
- Find a solution for get/setsockopt to no longer have to find options in Sets in order to apply the right type.