@ronkorving 2016-03-24
Some old code in C++ land can be removed, but I would rather do that post-review.