@reqshark 2016-03-24
looks like some of the older versions of libzmq aren't compatible with the changes at this point