@ronkorving 2016-03-24
Interesting. Travis was very slow to start working so I didn't get the feedback I was hoping for immediately. Anyway, it's midnight now so will look into it tomorrow.