@ronkorving 2016-03-25
For some reason, the event poller (UV_PollCallback) is not being triggered at all with my req/rep sockets. I'm a bit lost as to why.