@kkoopa 2016-03-25
For zmq2, you must check `ZMQ_EVENTS` after each call to `zmq_send` and `zmq_recv`. Easiest way is to insert calls to `socket->CallbackIfReady()`.