@ronkorving 2016-03-25
You're a lifesaver!

By the way, I did just check enabling UV_WRITABLE performance, and while it gives better throughput, latency increases. So I'm gonna keep it disabled for now (and hope your advice here works). We can gather more information on what is a good polling strategy another day.