@ronkorving 2016-03-25
What about zmq_sendmsg? This is breaking on ZMQ3 as well. Not just on 2.