@kkoopa 2016-03-25
For me it worked fine with the version of zmq3 I had installed. If you have a version it does not work on, I'll assume the same applies as for zmq2, at least until that point.

You should only ever check for `UV_READABLE`, anything else is wrong, that's the way the signaling mechanism is intended. The problem here is mostly that zmq has / had many warts and design flaws. Regarding this particular issue, I think it is buried somewhere in the documentation. The "file descriptor" returned via `ZMQ_FD` is not really a file descriptor, but it uses a similar interface. In zmq2 it amounts to a counter for events, when it is non-zero `uv_poll` detects it and now you need to poll again via `zmq_poll` because it is edge triggered. Now, whenever you call `zmq_send` or `zmq_recv`, that counter gets reset to 0, and so the fd-poller will not detect anything.