@kkoopa 2016-03-25
But it seems I might have had zmq v4 installed through the package named libzmq3... :confounded: 
```sh
libzmq3-dev | 4.0.5+dfsg-2+deb8u1build0.15.04.1
```