@ronkorving 2016-03-25
Oh god, the addition of these checks bring performance back to old numbers (40-50k msg/sec). But at least we shouldn't need this in ZMQ 4, right?

Still makes me think there's an incredible flaw in how I do reads (because I read all messages until the last ZMQ_RCVMORE.