@kkoopa 2016-03-25
zmq 4 seemed to work, so I guess it is not needed there. Also, it might be that you do not need to go through `zmq_poll`. A `getsockopt` for EVENTS and checking that might suffice, but it is essentially the same as `zmq_poll` with 0 timeout.

I would suspect the performance killer is the call from c++ to javascript, which happens for each message (batch). By doing less transitions from C++ to JavaScript, performance should improve.