@ronkorving 2016-03-25
Related question: is this a good time to drop ZMQ 2/3 support? :)