@ronkorving 2016-03-25
For some strange reason, the routing error test started to fail on me on ZMQ3, but I can look into that later. In any case, this mostly works now.