@kkoopa 2016-03-25
That's mostly just hiding the problem. You could try an ugly hack and change the zmq file descriptor back to nonzero after send or recv. Then, the uv poller would detect the change and do the rest.
The point here is to trade latency for performance, since I assume libuv polls at some fixed rate. Lower poll frequency means less calls to javascript, which means higher throughput.