@ronkorving 2016-03-25
Sorry, what exactly is hiding the problem? The `IsReady` calls over the `CallbackIfReady` calls? In other words, you're saying we need to trigger `onReady` in JS? Because right now, the readv and sendv functions loop to collect all message parts of a single message. That's where I get a lot of performance from. Is this inherently wrong?