@kkoopa 2016-03-25
I was referring to dropping support for zmq 2 and 3. The reason 4 works better is because there are less calls to javascript, since they can only happen when uv_poll detects a change on the file descriptor, as opposed to after every send and recv.

On March 25, 2016 2:58:02 PM GMT+02:00, Ron Korving <notifications@github.com> wrote:
>Sorry, what exactly is hiding the problem? The `IsReady` calls over the
>`CallbackIfReady` calls? In other words, you're saying we need to
>trigger `onReady` in JS? Because right now, the readv and sendv
>functions loop to collect all message parts of a single message. That's
>where I get a lot of performance from. Is this inherently wrong?
>
>---
>You are receiving this because you were mentioned.
>Reply to this email directly or view it on GitHub:
>https://github.com/JustinTulloss/zeromq.node/pull/499#issuecomment-201268581