@ronkorving 2016-03-25
Oh sure, dropping 2 and 3 is *not* a solution to the problems I had. But you can imagine my frustration caused me to toy with the idea ;) And thus, a side-question (outside of scope for this PR) is: how much longer do we wanna support 2 and 3?