@kkoopa 2016-03-25
Are they still being actively maintained by zmq?

On March 25, 2016 3:02:58 PM GMT+02:00, Ron Korving <notifications@github.com> wrote:
>Oh sure, dropping 2 and 3 is *not* a solution to the problems I had.
>But you can imagine my frustration caused me to toy with the idea ;)
>And thus, a side-question (outside of scope for this PR) is: how much
>longer do we wanna support 2 and 3?
>
>---
>You are receiving this because you were mentioned.
>Reply to this email directly or view it on GitHub:
>https://github.com/JustinTulloss/zeromq.node/pull/499#issuecomment-201270026