@ronkorving 2016-03-25
Nope: https://raw.githubusercontent.com/zeromq/zeromq3-x/master/NEWS
2014.