@ronkorving 2016-03-25
So regarding the idea of pulling out all message parts in C++ as long as SNDMORE is set, do you think this will always work reliably?