@kkoopa 2016-03-25
I think there exists an implemention which could achieve that, but I do not know what it is. Maybe the current could do that reliably, but I don't know what zmq has to say about that. I know for a fact that some of its socket types are unusable.

tldr: Dunno, try and see.

On March 25, 2016 3:25:12 PM GMT+02:00, Ron Korving <notifications@github.com> wrote:
>So regarding the idea of pulling out all message parts in C++ as long
>as SNDMORE is set, do you think this will always work reliably?
>
>---
>You are receiving this because you were mentioned.
>Reply to this email directly or view it on GitHub:
>https://github.com/JustinTulloss/zeromq.node/pull/499#issuecomment-201274965