@kkoopa 2016-03-25
If the reason for low throughput is due to the overhead of constantly making callbacks from C++ to JavaScript, a possible way of getting more performance is setting up a local (unix) socket for communication between libzmq in C++ and JavaScript and sending messages back and forth over that.
Would not need to change more than the sending and receiving of messages part. Configuration and such could still be kept mostly as-is.