@ronkorving 2016-03-28
The guarantees I was worrying about do seem to be in place btw. On another note, I'm very annoyed with this need for polling (which is poorly documented, or at least, I can't find much about it), but it fixes the issue.