@ronkorving 2016-03-28
I've also noticed (outside of the scope of this PR) some places where getsockopt doesn't check for EINTR. Perhaps most notably, the GetSockOpt method that is exposed to JS for strings (`Socket::GetSockOpt<char*>(int option)`).