@ronkorving 2016-04-01
@kkoopa Mind having another look? Travis fails but is unrelated. It just keeps failing on downloading libsodium (no idea why).